backend/requirements.txt
flask
flask-cors
easyocr
pillow
pdf2image
torch
